Create JSON to send to endpoint, import JSON of old evals
Implemented the creation of surveys and formatting them in the correct form to send to the API.

Takes in JSONs of the format that the API sends and able to load them in to the new form.  Currently hard coded in an example JSON which is automatically loaded

Endpoints are not working, servers are returning empty data?